Clearly denote adapter interface for file reloading
